Creating a Qml wrapper for QDesktopWebView.
The Qml wrapper for QDesktopWebView was named DeclarativeDesktopWebView.
PageGraphicsView class was removed and a new one named MainView that inherits
form QDeclarativeView was added. The idea is that MainView will grow until it
contains the whole UI.

To create a plugin for the new Qml elements will be a future step.