Use QML tabwidget instead of QTabWidget.
So we can get rid of almost all QWidget dependencies. What doesn't work
today is the spinner and the drag and drop for the tabs. The implementation
probably contains some bugs but we'll see :D.

I plan to rename PageWidget.cpp to another name in the next commit.
This name doesn't make sense but I wanted to keep the diff readable.

Reviewed-by: Luiz Agostini