Added support to triple click event for UrlBar.
TripleClickMonitor was added to act as an event filter on top of
QML's TextInput, as it currently doesn't have support for triple
clicks. The user expects that such event selects the entire text.

Once this filter detects a double click, if a new click event happens
immediately next, it will trigger the triple click.

Reviewed-by: Alexis Menard