Implement drag and drop support for tabs.
The geometry is not managed by anchors, it's on purpose because the
anchoring system allows only to anchor siblings. The idea here is to have
a parent object tab, the actual header that we drag, and the content which
is an empty item just used to calculate geometries because that one is
actually decoupled from the header we are dragging making possible to
reorganize the tabs without caring of the header.

Reviewed-by: Caio Oliveira