TabWidget: use toplevel id to refer to itself
Inside TabWidget.qml references to 'tabWidget' work by accident. When
the functions there run, we happen to have in the environment an item
named tabWidget. If I rename it to 'theTabWidget', it won't properly
work.

The usage of 'this' is also a little bit trick: it works sometimes, but
if we call the function without qualifying, like in C++ calling a method
on the same object, 'this' is undefined. To avoid later bugs, I prefer
to avoid it.

In both cases, the clean solution is to provide an id to the toplevel
of TabWidget.qml and use it.

Reviewed-by: Alexis Menard