Create a BrowserObject to expose functionality to QML
This commit adds a BrowserObject object, that will expose functionality
to QML, rather than making C++ peeking on QML to see what it needs to do.
The name is "long/unusual" to avoid clashing with QML ids.

The first property exposed is the windowTitle, so we can do the
controlling logic in QML world. In this commit we use a Binding {} object
to connect the windowTitle with the active tab title.

Our previous approach was to make C++ listen to the right titleChanged
and change itself. And there was a bug on it, if one page in a tab not
active changed the title (via document.title = "hahaha"), it would change
the title of the window.

Reviewed-by: Alexis Menard