Fix the tab title possibly higher than the tab.
If the user has a very big font the title of the tab could be higher than
the tab. We have two choices here, either we make the tabs growing
in height either we enforce the size just like Chrome. Today we don't have
resolution independent assets therefore we can't make the height of the tab
to resize so I pick enforcing a pixelSize for the font.

Reviewed-by: Caio Oliveira