Make "foo bar" urls to fallback to google.
When the user types a url like "foo" QUrl::fromUserInput return
true but then WebKit will fail and we fallback to google but if the
url of the user is "foo bar" QUrl::fromUserInput returns an invalid
url that we used to pass blindly to WebKit which of course was doing nothing.
If the conversion fails then we fallback to google.

Reviewed-by: Ademar Reis