Make sure we have a common appearance on all platforms.
We enforce the font family as well as the font size in snowshoe because
we don't have assets that could expand/shrink depending on the platform.
For now it is Arial, with a pixel size of 12. It could be ajusted later.
The url progress bar is also now looking like the mockups. The border
properties were not set correctly. 3px match to the image borders itself
and it also fix the bottom margin also not looking like the mockups.

Reviewed-by: Caio Oliveira