Make the databasemanager a singleton.
It is easier to use as we just instanciate it in the main.cpp and we
can use it no matter which window, no matter which popup. I believe
it is easier than passing pointers around. I also don't see the benefit
of multiple database manager as we will be probably have one model per
information we have, e.g. the bookmark model is sharable around.

Reviewed-by: Ademar Reis