Make the popup translucent, and add an API to get the maximum height and width the popup can expand.
We make it translucent so that any nice background with shadow can look
nice. I added an API to get the height and width of the screen so that the
users of PopupMenu can react to it and add for example a scroll bar/scroll buttons.

As a side note I fixed a const missing on one getter.

Reviewed-by: Ademar Reis