Workaround to fix the url bar being ugly.
The bug has been reported to the QML team but in order to make it
usable let's workaround it. The problem is that we set the y
of the parent of the UrlEdit to verticalMargins/2 bringing the result
3.5. The .5 is the problem. Visually it makes almost no different as
.5 is really hard to see so we can keep this value like this.

Bug is https://bugreports.qt.nokia.com/browse/QTBUG-20746 for the record.

Reviewed-by: Ademar Reis