Support mouse shortcuts to open link in new tab
With Middle click or Control + Left click, the link clicked will be
open in a new tab. It uses the new DesktopWebView API (WebKit r93427)
to intercept the navigation from the user and decide whether it should
continue (Use) or be ignored.

Given the proper mouse/keyboard state, we open a new tab, load the URL
there and ignore the original navigation.

This commit also modifies addNewTab() to return the created tab, so that
we can manipulate it.

Reviewed-by: Alexis Menard