Make MainView less dependent of QML structure
To get the WebView from a given Tab we were peeking into the UrlEdit,
which is a detail of our navigation bar. This commit changes the code
to avoid that and get the webview from higher level elements.

I've also renamed the function to describe what it was actually
doing. It wasn't opening a URL in a new tab but opening a URL in the
current active tab!

Note that this function isn't used for existing Ctrl-Click link to
open in new tab.

Reviewed-by: Alexis Menard