Restore the previously opened tabs
Snowshoe now saves and restores the tabs opened between
executions. Restored tabs are opened before URLs passed in the command
line.

Whenever a URL is changed, a tab is closed, or the tab order changes
our StateTracker is updated. When the application leaves, it persists
the list into our configuration file.

Things to improve:

- We could try to communicate the differences to the state tracker instead
  of the whole list of URLs everytime. But I have the intuition that this won't
  bring us much.

- It still doesn't take into account the case of the browser crashing,
  right now we only persist when the browser is exiting.

Reviewed-by: Rafael Brandão