Give TabWidget the geometry of the tab bar area
Now that the pages aren't inside the TabWidget, it can have the
geometry of the tab bar, and we can use it to anchor the content area.

It also means that TabWidget doesn't make sense as a toplevel anymore,
so I'm reverting it to be an item inside our toplevel.

Reviewed-by: Ademar Reis