Merge BrowserObject back into BrowserWindow
The indirection of having a separate object to expose to QML wasn't
buying us anything and BrowserWindow was being exposed already for other
usages. To make it even clearer, change the name that is exposed to QML
to be 'BrowserWindow' too, since 'View' is too general.

The original goal of BrowserObject is still fulfilled, we expose
things (now BrowserWindow itself) to QML instead of peeking into the toplevel
item.

Reviewed-by: Rafael Brandão