Make the popup for extra bookmarks work again
It wasn't properly working before. The popup wasn't being shown.

Instead of creating a new QQuickView (which brings its own QQuickCanvas) we now
do as QtQuick.Window experiments and create a new QQuickCanvas. This allows us
to interact more easily with the code to be loaded, by setting properties and using
JavaScript code.

This commit also tries to put the "action" as higherlevel as possible,
so now the Delegate doesn't call loadUrl() directly, but emits a signal, then the
DropDownMenuBookmark emits a signal which is handled by BookmarkBar. The reason
was to make code a bit less coupled.

The popup isn't translucent yet.  There are some other FIXMEs with further
improvements we can propose to Qt5, too.

Reviewed-by: Alexis Menard