Opening animation, part 1
Creates a new state for the "splash" opening. We need to add Luna logo
here in the future.  This commit also changes panelToggle and plusButton
to make use of anchors when possible, including in animations.

Changed the FavoritesPanel to allow triggering the opening animation for
the items instead of doing in the component completion. Some minor
renamings and the usage of PropertyAction instead of PropertyAnimation
with duration 0.

Opening animation should be visible from the device now!

Reviewed-by: NOBODY