Typing url should move user to "type-new-url" view
To type the url of a new tab we first create the tab with empty url
and then we move to the typing view to change the url of current tab.
Once you click into the url bar of an existing loaded page, then the
typing view is presented with the current url typed into text edit.
RootPage reference is propagated until StatusBarIndicator, so it can
trigger the change of view bottom-up.

Reviewed-by: Alexis Menard