Move code to position NavigationBar into NavigationPanel
The rationale to move the code is that NavigationPanel (and its
layouting code) is the responsible for positioning its children. We
later could even have the benefit of using anchoring between navigation bar,
current webview and tab bar.

NavigationBar is just a navigation bar, and if its position change
(imagine we remove the animation and make it always visible), the
NavigationBar itself won't change.

Reviewed-by: Rafael Brandão