Add the NavigationOverlay to our NavigationPanel
We are moving to a different approach to handle the interaction with
Snowshoe. When browsing the web, we let as much as possible screen
available for the web content, except for a bar with the tabs indicator.

When the current visible WebView is loading something, the tab
indicator grows showing an UrlBar. This UrlBar disappear two seconds
after the loading stops. This is controlled by
navigationBarHidingTimer.

If the user clicks the tabs indicator (either when loading or not
loading) we'll open not only the UrlBar but also put an Overlay on top
of web content.  The web content will be a "background" (not
interactive) and we'll use the whole area for gestures that change
tab, add tab, and other secondary actions.

The rationale summarized is that if the user clicked the tabs
indicator, then the user intends to interact with Snowshoe (manage
tabs, close tabs, etc), and not the web content. Clicking again on
the area brings the web content back.

To make the patch less confusing, I avoided some renamings that would
make the code more readable. They will come in later patches.
There is a new set of images coming as well, until they arrive we keep
the old ones.

Reviewed-by: Alexis Menard