Make SnowshoeWebView.url reflect the url requested to be loaded
We keep it as a buffer of what has been requested, then we commit that
change to the internal webview. Once the load has succeeded, the requested
url is then changed to the final url, as it could have been redirected.

Also, add entries on history only if it is not "about:blank" url, as it
could be a 404 page, for example.

Reviewed-by: Caio Oliveira
             Marcelo Lira