Workaround to avoid redirecting clicks to the webView on N9
N9's webView component captures all the events when active. To avoid
sending unneeded events to it (like when clicking on tabbar indicators)
the  webview's height was set according to the tabbar+navbar height. Also
the animations values where changed to avoid flickering. Opacity
property for overlay panel had to be used with an animation to avoid
flickering on N9.

Reviewed-by: trust me