Hide status bar until Qt5 orientation is fixed
QtComponents for QtQuick2 uses QWindow->reportContentOrientationChange
to tell the system that its contents will be on a different
orientation. It'll call QPlatformWindow->handleContentOrientationChange
which is a placeholder function. Until it is properly handled the
status bar will be being drawn twice for portrait apps as the system
will think the application is still in landscape mode.

There is a tentative fix at the link below

https://codereview.qt-project.org/#change,24423

Reviewed-by: Luciano Wolf