Revert hack to fix the status bar.
The content's orientation was fixed on qtbase but
it still does not show when not swiping on portrait mode,
appearing just its empty black space. When swiping the
system status bar appears correctly.

Reverting the hack at least fixes the screen dimension
problems that appeared when it was hidden.

Also updated BrowserWindowMobile to report its orientation
following the new orientation API.

Reviewed-by: Rafael Brandão