TabsModel::remove should not emit a dataChanged signal.
A dataChanged signal is not meant to be used when removing a row,
because the inner workings of a model expect to operate over a
still existing row, and insert it back on the model.

Without this patch you can see Snowshoe crashing by opening 5 tabs
(or any number whose modulo by 4 is 1; 4 is not magical, just the
number of tabs per page on the paged grid) and closing one.

Reviewed-by: Rafael Brandão