Avoid PageFillGrid to segfault when a page is removed
The problem was reproduced when you had 5 items (2 pages)
and removed one item from the first page. Not sure the exact
source of the problem but it seems to be Qt5-related (I'll try
to reduce into a small test case). As this bug is critical, it's
better to land as is and figure out the bug later.

The fix is to change our Repeater to load the maximum number of
pages, rather than creating as we need. I've also removed the
previous behavior of empty slots and now all the pages are created
once with empty slots and this layer is put behind the occupied slots.

Reviewed-by: Marcelo Lira