Activate QStyle theme for the desktop version.
There is no way to make custom scrollbars on the UIProcess side because
the way the flickable is handled in desktop mode. View = Content so we
can't get infos in QML about scrolling as there is no scrolling possible.
The mobile RenderTheme  doesn't have an implementation for the scrollbar
(as it assumes you can do it in the UIProcess) so we can fallback to
the QStyle RenderTheme. I think it is fine for snowshoe desktop to use it,
it is anyway way better than the mobile one (IMHO).

Reviewed-by: Caio Oliveira