Moved similar code to a new component PageThumbUrl.
The delegates of TabsPanel and TopSitesPanel share a lot of things
in common, most of those things are those pesky little details like
margin, font, color, and even an identical regular expression.
Stuff like that should be in the same place.

In addition, the margin and font problems of the thumbnail url texts
reported in the design review, were fixed.

Reviewed-by: Hugo Parente Lima