BookmarkModel::insert uses a more low level approach to data insertion.
It now uses a SQL query instead of a QSqlRecord. The reason for
using the former is that insertion through QSqlRecord increases the
row count, and emits a signal advertising it, before the row is
commited to the database. This inconsistent state is causing trouble
with the pin/unpin button in the mobile UI.

Reviewed-by: Rafael Brandão