Enumerate tabs correctly
As we use Repeater for each page, we get an index in the range from zero
to the number of items per page minus one for each component. By adding
the offset to the immediate parent, it becomes accessible inside the
delegate and then we can add the current page offset expected.

Reviewed-by: Caio Oliveira