Fix saving the opened urls when closing snowshoe
Summary:
There are three ways we can close Snowshoe

1) Closing the window from the Window Manager.
2) Using Ctrl-Shift-Q shortcut.
3) Using Ctrl-W to close last tab.

The cases (2) and (3) were quitting and leaking the BrowserWindow, so
ApplicationStateTracker destructor would not get called, and the tabs
opened were not persisted. Fixed this by closing the window instead of
quitting directly, so other cases work like case (1).

However QWindow::close() is not enough. I've created a bug for it, but
meanwhile we workaround by having our own close method.

Reviewers: darktears

Reviewed By: darktears

Differential Revision: http://review.qtlabs.org.br/D5